requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
loguru==0.6.0
lxml==4.9.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.2
openai==0.27.4
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==2.2.1
platformdirs==3.0.0
prompt-toolkit==3.0.37
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyspnego==0.9.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-ntlm==1.2.0
requests-toolbelt==1.0.0
SharePlum==0.5.1
six==1.16.0
stack-data==0.6.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
win32-setctime==1.1.0
yarl==1.8.2
